# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TCG Tools",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/premiacao/presets/{preset_id}": {
      "get": {
        "tags": [
          "premiacao"
        ],
        "summary": "Get Preset",
        "operationId": "get_preset_api_v1_premiacao_presets__preset_id__get",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "premiacao"
        ],
        "summary": "Update Preset",
        "operationId": "update_preset_api_v1_premiacao_presets__preset_id__put",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          },
          {
            "name": "X-Presets-Mtime",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Presets-Mtime"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/premiacao/calcular": {
      "post": {
        "tags": [
          "premiacao"
        ],
        "summary": "Calcular Endpoint",
        "operationId": "calcular_endpoint_api_v1_premiacao_calcular_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalcularRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalcularResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/premiacao/tabela": {
      "get": {
        "tags": [
          "premiacao"
        ],
        "summary": "Tabela",
        "operationId": "tabela_api_v1_premiacao_tabela_get",
        "parameters": [
          {
            "name": "ate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Ate"
            }
          },
          {
            "name": "preset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TabelaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/premiacao/export": {
      "post": {
        "tags": [
          "premiacao"
        ],
        "summary": "Export Csv",
        "operationId": "export_csv_api_v1_premiacao_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CalcularRequest": {
        "properties": {
          "jogadores": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Jogadores"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "valor_inscricao": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Inscricao"
          }
        },
        "type": "object",
        "required": [
          "jogadores"
        ],
        "title": "CalcularRequest"
      },
      "CalcularResponse": {
        "properties": {
          "jogadores": {
            "type": "integer",
            "title": "Jogadores"
          },
          "premiados": {
            "type": "integer",
            "title": "Premiados"
          },
          "premios": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Premios"
          },
          "total_inscricoes": {
            "type": "number",
            "title": "Total Inscricoes"
          },
          "creditos": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creditos"
          }
        },
        "type": "object",
        "required": [
          "jogadores",
          "premiados",
          "premios",
          "total_inscricoes"
        ],
        "title": "CalcularResponse"
      },
...
      "ExportRequest": {
        "properties": {
          "ate": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Ate"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          }
        },
        "type": "object",
        "required": [
          "ate"
        ],
        "title": "ExportRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PresetBody": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "min_jogadores": {
            "type": "integer",
            "title": "Min Jogadores"
          },
          "min_premiados": {
            "type": "integer",
            "title": "Min Premiados"
          },
          "max_premiados": {
            "type": "integer",
            "title": "Max Premiados"
          },
          "crescimento": {
            "type": "integer",
            "title": "Crescimento"
          },
          "r": {
            "type": "number",
            "title": "R"
          },
          "casas_decimais": {
            "type": "integer",
            "title": "Casas Decimais"
          }
        },
        "type": "object",
        "required": [
          "label",
          "min_jogadores",
          "min_premiados",
          "max_premiados",
          "crescimento",
          "r",
          "casas_decimais"
        ],
        "title": "PresetBody"
      },
...
      "TabelaLinha": {
        "properties": {
          "jogadores": {
            "type": "integer",
            "title": "Jogadores"
          },
          "premiados": {
            "type": "integer",
            "title": "Premiados"
          },
          "premios": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Premios"
          }
        },
        "type": "object",
        "required": [
          "jogadores",
          "premiados",
          "premios"
        ],
        "title": "TabelaLinha"
      },
      "TabelaResponse": {
        "properties": {
          "linhas": {
            "items": {
              "$ref": "#/components/schemas/TabelaLinha"
            },
            "type": "array",
            "title": "Linhas"
          }
        },
        "type": "object",
        "required": [
          "linhas"
        ],
        "title": "TabelaResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}